✨ feat: move useGetWebsocketUrlQuery for websocket URL retrieval; update imports and add Google provider to RealtimeVoiceProviders enum
