+++ client/src/data-provider/queries.ts
@@ -532,3 +532,18 @@ export const useUserTermsQuery = (
     ...config,
   });
 };
+
+export const useGetWebsocketUrlQuery = (
+  config?: UseQueryOptions<t.TWebsocketUrlResponse>,
+): QueryObserverResult<t.TWebsocketUrlResponse> => {
+  return useQuery<t.TWebsocketUrlResponse>(
+    [QueryKeys.websocketUrl],
+    () => dataService.getWebsocketUrl(),
+    {
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      refetchOnMount: false,
+      ...config,
+    },
+  );
+};
